+++ src/hooks/utils/useCanUserSubmitProposal.ts
@@ -2,12 +2,12 @@ import { abis } from '@fractal-framework/fractal-contracts';
 import { useCallback, useEffect, useState } from 'react';
 import { Address, getContract } from 'viem';
 import { useAccount, usePublicClient } from 'wagmi';
+import { FeatureFlag, isFeatureEnabled } from '../../helpers/featureFlags';
 import { useFractal } from '../../providers/App/AppProvider';
 import { useSafeAPI } from '../../providers/App/hooks/useSafeAPI';
 import { useDaoInfoStore } from '../../store/daoInfo/useDaoInfoStore';
 import { GovernanceType } from '../../types';
 import useVotingStrategiesAddresses from './useVotingStrategiesAddresses';
-import { isDemoMode } from '../../helpers/featureFlags';
 
 export function useCanUserCreateProposal() {
   const {
@@ -117,7 +117,8 @@ export function useCanUserCreateProposal() {
 
   useEffect(() => {
     const loadCanUserCreateProposal = async () => {
-      const newCanCreateProposal = isDemoMode() || (await getCanUserCreateProposal());
+      const newCanCreateProposal =
+        isFeatureEnabled(FeatureFlag.demoMode) || (await getCanUserCreateProposal());
       if (newCanCreateProposal !== canUserCreateProposal) {
         setCanUserCreateProposal(newCanCreateProposal);
       }
