+++ src/main.tsx
@@ -6,7 +6,8 @@ import './assets/css/Markdown.css';
 import './assets/css/sentry.css';
 import './assets/css/Toast.css';
 import './insights';
-import { EnvironmentFeatureFlags, FeatureFlags } from './helpers/featureFlags';
+import { EnvironmentFeatureFlags } from './helpers/environmentFeatureFlags';
+import { FeatureFlags } from './helpers/featureFlags';
 import { runMigrations } from './hooks/utils/cache/runMigrations';
 import { useNetworkConfigStore } from './providers/NetworkConfig/useNetworkConfigStore';
 import Providers from './providers/Providers';
